refactor: rename load_new_metadata to load_metadata
Remove "new" suffix now that legacy load_metadata.py no longer exists.
Update all imports, test files, and spec documentation.

Files renamed:
- aeon/dj_pipeline/utils/load_new_metadata.py → load_metadata.py
- tests/.../test_load_new_metadata_unit.py → test_load_metadata_unit.py
- tests/.../test_load_new_metadata_integration.py → test_load_metadata_integration.py

Also adds uv.lock for reproducible builds.
