+++ aeon/dj_pipeline/utils/load_metadata.py
